+++ src/main/java/com/example/marluse/estoque/repository/ProdutoRepository.java
@@ -4,6 +4,7 @@ import com.example.marluse.estoque.model.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
@@ -14,7 +15,7 @@ public interface ProdutoRepository extends JpaRepository<Produto, String> {
 
     List<Produto> findByAtivoTrue();
 
-    List<Produto> findByQuantidadeEstoqueLessThanEqualAndAtivoTrue(Integer quantidade);
+    List<Produto> findByQuantidadeEstoqueLessThanEqualAndAtivoTrue(BigDecimal quantidade);
 
     @Query("SELECT p FROM Produto p WHERE p.quantidadeEstoque <= p.estoqueMinimo AND p.ativo = true AND p.rascunho = false")
     List<Produto> findEstoqueBaixo();
