+++ src/main/java/com/example/marluse/relatorios/dto/TopProdutoResponse.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public record TopProdutoResponse(
         String nome,
-        long quantidade,
+        BigDecimal quantidade,
         BigDecimal lucro,
         BigDecimal custo,
         BigDecimal total
